Move AliasRegistration and ChangeDetection to new Installation namespace
