+++ developer-cli/Program.cs
@@ -1,8 +1,8 @@
-﻿using PlatformPlatform.DeveloperCli;
+﻿using PlatformPlatform.DeveloperCli.Installation;
 using Spectre.Console;
 
 ChangeDetection.EnsureCliIsCompiledWithLatestChanges(args);
-Installation.EnsureAliasIsRegistered();
+AliasRegistration.EnsureAliasIsRegistered();
 
 var command = args.FirstOrDefault();
 AnsiConsole.MarkupLine($"Unknown command [green]{command}[/]. Use [green]--help[/] to see available commands.");
